+++ Interfaces/Orders/IOrderRepo.cs
@@ -7,8 +7,8 @@ namespace MediNet_BE.Interfaces.Orders
     {
         public Task<List<OrderDto>> GetAllOrderAsync();
         public Task<OrderDto> GetOrderByIdAsync(int id);
-        public Task<Order> AddOrderAsync(OrderDto orderDto);
-        public Task UpdateOrderAsync(OrderDto orderDto);
+        public Task<Order> AddOrderAsync(OrderCreateDto orderDto);
+        public Task UpdateOrderAsync(OrderCreateDto orderDto);
         public Task DeleteOrderAsync(int id);
     }
 }
