+++ Interfaces/Orders/IProductRepo.cs
@@ -8,8 +8,8 @@ namespace MediNet_BE.Interfaces.Orders
     {
         public Task<List<ProductDto>> GetAllProductAsync();
         public Task<ProductDto> GetProductByIdAsync(int id);
-        public Task<Product> AddProductAsync(ProductDto productDto);
-        public Task UpdateProductAsync(ProductDto productDto);
+        public Task<Product> AddProductAsync(ProductCreateDto productDto);
+        public Task UpdateProductAsync(ProductCreateDto productDto);
         public Task DeleteProductAsync(int id);
     }
 }
